# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShadyCard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: Current user orders
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [all, wait, accept, reject] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id: { type: string }
        orderNumber: { type: string }
        productId: { type: string }
        productName: { type: string }
        productImage: { type: string }
        quantity: { type: number }
        userIdentifier: { type: string }
        totalUsd: { type: number }
        totalSyp: { type: number }
        status: { type: string, enum: [wait, accept, reject] }
        createdAt: { type: string }
      required: [id, orderNumber, productId, productName, quantity, totalUsd, totalSyp, status, createdAt]
    CreateOrderBody:
      type: object
      properties:
        productId: { type: string }
        quantity: { type: number }
        userIdentifier: { type: string }
      required: [productId, quantity]
...