# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShadyCard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  productId: string;
  quantity: number;
  userIdentifier?: string;
}